Initial commit, basic functionality
The code lays the foundations for feedback post enforcement, enforcing channel, character minimum and keywords.

Karma system - When a user posts a message that is a reply to a feedback post and the reply content qualifies as feedback. The user is awarded 1 karma which is appended to karma.json, a dictionary mapping user ids to thread ids.
The user can only receive 1 karma per thread.

Keyword system - keywords are enforced, mix feedback should contain key words which are common in music production and sound design.

Some basic logging has been added in the form of print statements, this will be improved later.,

- Included code to prevent feedback poster receiving karma from replies to their own thread